% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_occurrences.R
\name{average_occurences}
\alias{average_occurences}
\title{Request Message and Chat Data}
\usage{
average_occurences(message_data)
}
\arguments{
\item{message_data}{A tibble returned from palver::request_data or palver::request_messages}
}
\value{
The input tibble with aditional colunms on occurrences of the same text or midia and their average
}
\description{
Request Message and Chat Data
}
