% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_token}
\alias{get_token}
\title{Get Token for Palver Data API}
\usage{
get_token(email, password)
}
\arguments{
\item{email}{E-mail from your Palver account}

\item{password}{Password from your Palver account}
}
\value{
A token string to access Palver Data
}
\description{
Get Token for Palver Data API
}
