% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_data.R
\name{request_data}
\alias{request_data}
\title{Request Message and Chat Data}
\usage{
request_data(
  query,
  country = NULL,
  region = NULL,
  startDate,
  endDate,
  sentiment = NULL,
  source = c("whatsapp", "telegram"),
  is_spam = NULL,
  is_nsfw = NULL,
  is_news_related = NULL,
  is_potentially_misleading = NULL,
  lang = NULL,
  transcript_lang = NULL,
  ocr_lang = NULL,
  tags = NULL,
  type_label = NULL,
  sortField = c("datetime", "forwarding_score"),
  sortOrder = c("desc", "asc"),
  token
)
}
\arguments{
\item{query}{Query in lucene's syntax}

\item{country}{Country ISO code}

\item{region}{Brazilian region code, (only if country is BR)}

\item{startDate}{Start date in ISO format}

\item{endDate}{End date in ISO format}

\item{sentiment}{Sentiment of the message}

\item{source}{Source of the message}

\item{is_spam}{Include or exclude messages classified as spam}

\item{is_nsfw}{Include or exclude messages classified as NSFW}

\item{is_news_related}{Iclude or exclude messages that are related to news}

\item{is_potentially_misleading}{Include or exclude messages that contains news and are classified as potentially misleading}

\item{lang}{Language of the message}

\item{transcript_lang}{Language of the audio transcript}

\item{ocr_lang}{Language of the image OCR}

\item{tags}{Tags associated with the message's chat}

\item{type_label}{Type of content in message}

\item{sortField}{Sort Field}

\item{sortOrder}{Sort Order}

\item{token}{Object with your token from function get_token}
}
\value{
A tibble from data required
}
\description{
Request Message and Chat Data
}
