% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_media.R
\name{request_media}
\alias{request_media}
\title{Request Media}
\usage{
request_media(
  media_path,
  download = c(TRUE, FALSE),
  path,
  file_name = "media",
  source = c("whatsapp", "telegram", "reddit", "tiktok", "press", "news", "radio.medias",
    "television", "youtube", "twitter"),
  token
)
}
\arguments{
\item{media_path}{Media path from messages data}

\item{download}{Logical parameter for downloading}

\item{path}{Path file to save media}

\item{file_name}{File name to save media}

\item{source}{Source of messages data}

\item{token}{Token access}
}
\value{
Download media or return its URL download path
}
\description{
Request Media
}
