% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_chats.R
\name{request_chats}
\alias{request_chats}
\title{Request Chat Data}
\usage{
request_chats(
  query = NULL,
  id = NULL,
  name = NULL,
  source = c("whatsapp", "telegram", "reddit", "press", "news", "radio.medias",
    "youtube", "twitter"),
  token
)
}
\arguments{
\item{query}{Boolean query in lucene's syntax}

\item{id}{The group's unique identifier}

\item{name}{Exact group name}

\item{source}{Groups' origin platform}

\item{token}{Object with your token from function get_token}
}
\value{
A tibble from data required
}
\description{
Request Chat Data
}
